增加了上传和读取文件的功能

+++ Web_Chroma.py
@@ -9,6 +9,15 @@ import tempfile
 import split
 import requests
 import json
+from reportlab.lib.pagesizes import letter, landscape
+from reportlab.lib import colors
+from reportlab.lib.styles import getSampleStyleSheet
+from reportlab.platypus import SimpleDocTemplate, Paragraph
+from bs4 import BeautifulSoup
+import tempfile
+import PyPDF2
+import os
+# 创建 PDF 文件
 
 import os
 
@@ -18,30 +27,65 @@ class WebChroma:
         self.embedding_function = OpenAIEmbeddings(model="text-embedding-ada-002")
         self.api_key=os.getenv('coreapikey')
 
-    @staticmethod
-    def download_pdf(url):
+    # @staticmethod
+    # def download_pdf(url):
+    #     headers = {
+    #         'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36',
+    #     }
+    #     try:
+    #         response = requests.get(url,headers=headers)
+    #         if response.status_code == 200:
+    #
+    #             return response.content
+    #         else:
+    #             return None
+    #     except :
+    #         return None
+
+    def download(self,url):
         headers = {
             'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36',
+            'Authorization': f'Bearer {self.api_key}'  # 替换为您的API密钥
         }
         try:
-            response = requests.get(url,headers=headers)
+            response = requests.get(url, headers=headers)
             if response.status_code == 200:
+                # st.write(response.content)
                 return response.content
             else:
+                print("The downloaded content is not exist.")
                 return None
-        except :
+            # else:
+            #     print(f"Failed to download PDF. Status code: {response.status_code}")
+            #     return None
+        except Exception as e:
+            print(f"An error occurred: {e}")
             return None
-
     @staticmethod
     def read_pdf(pdf_content):
         # 创建一个临时文件，并将PDF内容写入其中
-        with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-            temp_file.write(pdf_content)
-            temp_file.seek(0)
-            pdf_reader = PyPDF2.PdfReader(temp_file.name)
-        # 删除临时文件
-        os.remove(temp_file.name)
-        return pdf_reader
+        # st.write(pdf_content)
+        if pdf_content[:4] == b"%PDF":
+            with tempfile.NamedTemporaryFile(delete=False) as temp_file:
+                temp_file.write(pdf_content)
+                temp_file.seek(0)
+                pdf_reader = PyPDF2.PdfReader(temp_file.name)
+            # 删除临时文件
+            os.remove(temp_file.name)
+            return pdf_reader
+        else:
+            return None
+
+    @staticmethod
+    def read_html(content):
+        if content.lstrip().startswith(b'<!DOCTYPE') or content.lstrip().startswith(b'<html'):
+            # 解码字节对象为字符串
+            html_content_as_str = content.decode('utf-8')
+            # 在这里，您可以选择将 html_content_as_str 保存到文件或数据库中
+            return html_content_as_str
+        else:
+            return None
+
 
     @staticmethod
     def load_all_stopwords(dir_path):
@@ -54,7 +98,7 @@ class WebChroma:
                 jieba.analyse.set_stop_words(file_path)
 
     def pdf_text(self,url:str):
-        pdf_content = self.download_pdf(url)
+        pdf_content = self.download(url)
         if pdf_content:
             # 创建PDF阅读器对象
             pdf_reader = self.read_pdf(pdf_content)
@@ -140,22 +184,8 @@ class WebChroma:
             return out_results[:limit]
         else:
             return None
-
-    def process_pdf(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=1000):
+    def pdf_content_web(self,pdf_reader,url,tags,title_year_authors='',MAX_TOKENS:int=1000):
         strings = []
-        self.load_all_stopwords('./stopwords-master')
-        title_year_authors = f'year:{year}，title:{title}，authors:{authors}'
-        if abstract is None or not isinstance(abstract, str):
-            abstract = ''
-        tags = jieba.analyse.extract_tags(abstract, topK=10)
-        st.caption((title, url))
-        pdf_content = self.download_pdf(url)
-        if pdf_content is None:
-            return
-        try:
-            pdf_reader = self.read_pdf(pdf_content)
-        except:
-            return
         try:
             for page in pdf_reader.pages:
                 all_text = page.extract_text().replace('...', '').replace('..', '').strip()
@@ -169,14 +199,58 @@ class WebChroma:
         wikipedia_strings = []
         for section in strings:
             wikipedia_strings.extend(split.split_strings_from_subsection_pdf(section, max_tokens=MAX_TOKENS))
-
         ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
         docsearch = Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
                                       persist_directory=self.vector_folder, ids=ids)
         return docsearch
-
-    from concurrent.futures import ThreadPoolExecutor
-    async def process_pdf_async(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=1000):
+    def html_content_web(self,html_reader,url,MAX_TOKENS:int=1000):
+        strings = []
+        strings.append(([f'{url}'], f'{html_reader}'))
+        MAX_TOKENS = MAX_TOKENS  # Adjust as needed
+        wikipedia_strings = []
+        for section in strings:
+            wikipedia_strings.extend(split.split_strings_from_subsection(section, max_tokens=MAX_TOKENS))
+        st.write(wikipedia_strings)
+        # print(wikipedia_strings)
+        ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
+        docsearch = Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
+                                      persist_directory=self.vector_folder, ids=ids)
+        return docsearch
+    # def process_pdf(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=1000):
+    #     strings = []
+    #     self.load_all_stopwords('./stopwords-master')
+    #     title_year_authors = f'year:{year}，title:{title}，authors:{authors}'
+    #     if abstract is None or not isinstance(abstract, str):
+    #         abstract = ''
+    #     tags = jieba.analyse.extract_tags(abstract, topK=10)
+    #     st.caption((title, url))
+    #     pdf_content = self.download(url)
+    #     if pdf_content is None:
+    #         return
+    #     try:
+    #         pdf_reader = self.read_pdf(pdf_content)
+    #     except:
+    #         return
+    #     try:
+    #         for page in pdf_reader.pages:
+    #             all_text = page.extract_text().replace('...', '').replace('..', '').strip()
+    #             formatted_url = f'url:{url}'
+    #             formatted_tags = f'abstract keyword:{tags}'
+    #             text = [all_text]
+    #             strings.append((title_year_authors, formatted_url, formatted_tags, text))
+    #     except UnicodeDecodeError:
+    #         return
+    #     MAX_TOKENS = MAX_TOKENS  # Adjust as needed
+    #     wikipedia_strings = []
+    #     for section in strings:
+    #         wikipedia_strings.extend(split.split_strings_from_subsection_pdf(section, max_tokens=MAX_TOKENS))
+    #
+    #
+    #     ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
+    #     docsearch = Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
+    #                                   persist_directory=self.vector_folder, ids=ids)
+    #     return docsearch
+    def process(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=1000):
         strings = []
         self.load_all_stopwords('./stopwords-master')
         title_year_authors = f'year:{year}，title:{title}，authors:{authors}'
@@ -184,31 +258,81 @@ class WebChroma:
             abstract = ''
         tags = jieba.analyse.extract_tags(abstract, topK=10)
         st.caption((title, url))
-        pdf_content = self.download_pdf(url)
+        pdf_content = self.download(url)
         if pdf_content is None:
             return
-        try:
-            pdf_reader = self.read_pdf(pdf_content)
-        except:
-            return
-        try:
-            for page in pdf_reader.pages:
-                all_text = page.extract_text().replace('...', '').replace('..', '').strip()
-                formatted_url = f'url:{url}'
-                formatted_tags = f'abstract keyword:{tags}'
-                text = [all_text]
-                strings.append((title_year_authors, formatted_url, formatted_tags, text))
-        except UnicodeDecodeError:
-            return
-        MAX_TOKENS = MAX_TOKENS  # Adjust as needed
-        wikipedia_strings = []
-        for section in strings:
-            wikipedia_strings.extend(split.split_strings_from_subsection_pdf(section, max_tokens=MAX_TOKENS))
+        pdf_reader = self.read_pdf(pdf_content)
+        if pdf_reader is not None:
+            docsearch = self.pdf_content_web(pdf_reader, url, tags, title_year_authors, MAX_TOKENS)
+        else:
+            html_reader = self.read_html(pdf_content)
+            if html_reader is not None:
+                docsearch = self.html_content_web(html_reader, url, MAX_TOKENS)
+            else:
+                docsearch = None  # 或者你可以设定其他的默认值，或抛出异常
+        return docsearch
 
-        ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
-        Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
-                                      persist_directory=self.vector_folder, ids=ids)
+    from concurrent.futures import ThreadPoolExecutor
+    # async def process_pdf_async(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=1000):
+    #     strings = []
+    #     self.load_all_stopwords('./stopwords-master')
+    #     title_year_authors = f'year:{year}，title:{title}，authors:{authors}'
+    #     if abstract is None or not isinstance(abstract, str):
+    #         abstract = ''
+    #     tags = jieba.analyse.extract_tags(abstract, topK=10)
+    #     st.caption((title, url))
+    #     pdf_content = self.download(url)
+    #     if pdf_content is None:
+    #         return
+    #     try:
+    #         pdf_reader = self.read_pdf(pdf_content)
+    #     except:
+    #         return
+    #     try:
+    #         for page in pdf_reader.pages:
+    #             all_text = page.extract_text().replace('...', '').replace('..', '').strip()
+    #             formatted_url = f'url:{url}'
+    #             formatted_tags = f'abstract keyword:{tags}'
+    #             text = [all_text]
+    #             strings.append((title_year_authors, formatted_url, formatted_tags, text))
+    #     except UnicodeDecodeError:
+    #         return
+    #     MAX_TOKENS = MAX_TOKENS  # Adjust as needed
+    #     wikipedia_strings = []
+    #     for section in strings:
+    #         wikipedia_strings.extend(split.split_strings_from_subsection_pdf(section, max_tokens=MAX_TOKENS))
+    #
+    #     ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
+    #     docsearch =Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
+    #                                   persist_directory=self.vector_folder, ids=ids)
+    #     return docsearch
 
+    # def process_html(self, url, title='', authors='', year='', abstract='',MAX_TOKENS:int=3000):
+    #     strings = []
+    #     self.load_all_stopwords('./stopwords-master')
+    #     title_year_authors = f'year:{year}，title:{title}，authors:{authors}'
+    #     if abstract is None or not isinstance(abstract, str):
+    #         abstract = ''
+    #     tags = jieba.analyse.extract_tags(abstract, topK=10)
+    #     st.caption((title, url))
+    #     pdf_content = self.download(url)
+    #     if pdf_content is None:
+    #         return
+    #     try:
+    #         html_reader=self.read_html(pdf_content)
+    #     except:
+    #         return
+    #     strings.append(([f'{url}'], f'{html_reader}'))
+    #     MAX_TOKENS = MAX_TOKENS  # Adjust as needed
+    #     wikipedia_strings = []
+    #     for section in strings:
+    #         wikipedia_strings.extend(split.split_strings_from_subsection(section, max_tokens=MAX_TOKENS))
+    #     st.write(wikipedia_strings)
+    #
+    #     ids = [f"{url}_{i}" for i in range(1, len(wikipedia_strings) + 1)]
+    #     docsearch = Chroma.from_texts(wikipedia_strings, self.embedding_function, collection_name="state-of-union",
+    #                                   persist_directory=self.vector_folder, ids=ids)
+    #     return docsearch
 
     def search_research_articles(self,query: str,search_query:str):
         st.write('正在下载论文，这个过程可能持续几分钟，请耐心等待。。。。')
@@ -222,27 +346,30 @@ class WebChroma:
             abstract = result['abstract']
             authors = result['authors']
             year = result['year']
-            docsearch=self.process_pdf(url,title,abstract,authors,year,2000)
-        docs = docsearch.similarity_search(search_query, k=7)
-        return docs
+            docsearch=self.process(url,title,abstract,authors,year,2000)
+        if docsearch is not None:
+            docs = docsearch.similarity_search(search_query, k=7)
+            return docs
+        else:
+            return"docsearch is None, may be you should download pdf by yourself ."
 
-    async def search_research_articles_async(self,query: str):
-        st.write('正在下载论文，这个过程可能持续几分钟，请耐心等待。。。。')
-        results = self.search_research_title_url_abstract(query)
-        if results is None:
-            return '没有找到论文，你可以去其他网址下载pdf后，然后进行分析'
-        tasks = []
-        for result in results:
-            strings = []
-            url = result['url']['url']
-            title = result['title']
-            abstract = result['abstract']
-            authors = result['authors']
-            year = result['year']
-            task = self.process_pdf_async(url, title, abstract, authors, year, 2000)
-            tasks.append(task)
-        docsearches = await asyncio.gather(*tasks)
-        return docsearches
+    # async def search_research_articles_async(self,query: str):
+    #     st.write('正在下载论文，这个过程可能持续几分钟，请耐心等待。。。。')
+    #     results = self.search_research_title_url_abstract(query)
+    #     if results is None:
+    #         return '没有找到论文，你可以去其他网址下载pdf后，然后进行分析'
+    #     tasks = []
+    #     for result in results:
+    #         strings = []
+    #         url = result['url']['url']
+    #         title = result['title']
+    #         abstract = result['abstract']
+    #         authors = result['authors']
+    #         year = result['year']
+    #         task = self.process_pdf_async(url, title, abstract, authors, year, 2000)
+    #         tasks.append(task)
+    #     docsearches = await asyncio.gather(*tasks)
+    #     return docsearches
 
     # async def read_research_articles(self,input_str: str):
     #     st.write('正在下载论文，这个过程可能持续几分钟，请耐心等待。。。。')
@@ -261,6 +388,7 @@ class WebChroma:
     #     # docs = docsearches.similarity_search(input_str, k=10)
     #     return docsearches
     def read_research_articles(self,input_url: str,search_query:str=''):
+        # global docsearch
         # global docsearch
         st.write('正在下载论文，这个过程可能持续几分钟，请耐心等待。。。。')
         if isinstance(input_url, list):
@@ -286,11 +414,15 @@ class WebChroma:
             abstract = ''
             authors = ''
             year = ''
-            docsearch = self.process_pdf(url, title, abstract, authors, year, 1000)
+            docsearch = self.process(url, title, abstract, authors, year, 2000)
         if search_query=='':
             search_query='content'
-        docs = docsearch.similarity_search(search_query, k=7)
-        return docs
+        if docsearch is not None:
+            docs = docsearch.similarity_search(search_query, k=3)
+            return docs
+        else:
+            return "docsearch is None, may be you should download pdf by yourself"
+
 
     def search_Cache(self,query: str):
         wikipedia_strings = []
@@ -346,5 +478,5 @@ if __name__ == "__main__":
 #     s=asyncio.run(chroma.search_research_articles_async('ai'))
 #     s=chroma.search_research_articles('ai')
 #     s=chroma.search_Cache('ai')
-    s=chroma.read_research_articles("['https://core.ac.uk/download/20659569.pdf']",'ai')
+    s=chroma.read_research_articles("['https://core.ac.uk/download/389320849.pdf']",'')
     print(s)
